### 2023-04-12 #5359 @oliviaguest
@peilun-he thanks for choosing JOSS. Can you elaborate on what has changed since this: https://github.com/openjournals/joss-reviews/issues/5172? Thanks!
