### 2023-04-15 #5359 @peilun-he
@oliviaguest Dear Olivia, we would like you to kindly consider this submission of publication to JOSS. We would like to point out the contributions of this application in R and the associated paper:
 1. There are no existing polynomial diffusion simulation and estimation packages. The first contribution is therefore to build a library in R that both simulates the discrete time polynomial diffusion models (multivariate), as well as forms estimate technique for both the likelihood and state estimation in this class of non-linear non-Gaussian models. In our opinion, this constitutes a non-trivial technique implementation of recently created diffusion models of the relevance of financial mathematics.
2. The second contribution was to develop stochastic pricing model for futures term structures using this class of polynomial diffusion models. The generalises classical existing stochastic models widely used in commodity futures modelling. This was also implemented as a part of the R implementation.
3. The third contribution was to develop a user-friendly GUI so that users would visualise the outputs of the simulation and estimation results obtained from the application created.
4. Lastly, for user efficiency, we also provided the ability to export the results to a csv for additional interrogation should that be desirable.

Please note that the development of the code to perform this estimation task took one and a half years to develop and test. There are approximately over 1000 lines of code in R. Many of these are written by the research team and not relying on simply calling other packages. Therefore, we would like to point out that whilst the Github repository does not demonstrate a long history of user commits, this is representing the amount of effort required to code this package. The majority of the development was done outside the Github repository. We mention this since we submitted this project to JOSS and we received no review due to the superficial assessment that the commit history is too short and the application had a GUI and that perhaps therefore there was no substantial contribution. We would like to dispute this assessment and we kindly request your reconsideration of this work, as we believe the work presented is both substantial, novel, and timely for the computational finance community. We hope you will therefore please consider this work, and we look forward to receiving your feedback.
