### 2023-05-11 #5359 @samhforbes
Hi @peilun-he
I'm currently looking at your submission, as well as the reasons for the previous rejection.
At the current stage, this submission falls short of JOSS requirements, but I also recognise the contributions the software makes. My concerns are as follows:
1. Documentation. The README does not address how to launch the app, and the documentation is scarce as to what changing certain values in the app mean. We expect software to be useable by people who are not the developers, but some documentation is necessary for this. Given it is an R package you could consider creating a vignette. There should also be relevant citations of publications in that documentation.
2. Can you consider (or justify why not) including a web interface containing the app so that each user does not need to build it locally? The shiny apps webpage has limited plans for example, and this would improve accessibility.
3. Would you consider packaging this into an R package rather than a script with helpers? This is standard in R development, and while not necessarily a deal-breaker, could reasonably be expected by users, unless there is a good reason not to.

Depending on if / how these concerns are addressed I can then consider sending this software out for review, but addressing these concerns does not at this stage guarantee a positive outcome.
