### 2023-06-14 #5359 @peilun-he
Hi @samhforbes, after discussing with all co-authors, we have made the following changes:
1. The app is deployed on a Shiny server. Users can run the app by going to https://peilunhe.shinyapps.io/pdsim/ (the link is also provided in README file).
2. This app is packaged into an R package so users can also download it by using the "devtools::install_github()" function and run it on their local machine.
3. User guide and how to install the app are added into the README file.
All these changes have been merged into the main branch. Thank you very much for your patience.
