### 2023-06-14 #5359 @samhforbes
Hi @peilun-he
Thanks for making these changes. I do think this has substantially improved the work you've submitted.
I still want to push on the first point I raised.
Specifically: I see much improved documentation, but there needs to be documentation on what the values in the app represent and what a change in those values represent. So something like a worked example where you show how to input it, or really clear vignettes (I noticed that's not populated yet so it might be on the way?) would be really helpful.
Other than this I'm happy with the changes you've made, and if you are willing to add this level of documentation, then I would be happy to send it out for peer review.
Is that something you would be willing to do?
