### 2023-06-22 #5359 @peilun-he
Hi @samhforbes
I have added a vignette to the package, which includes some GIFs about how to use this app. The meanings of some values are explained by the GIFs, and others are defined in the user guide (which is available in both the README file and the app).
