### 2023-08-16 #5762 @editorialbot
[REVIEW]: PDSim: A Shiny App for Polynomial Diffusion Model Simulation and Estimation
**Submitting author:** @peilun-he (<a href="http://orcid.org/0000-0002-2740-3390">Peilun He</a>)
**Repository:** https://github.com/peilun-he/polynomial-diffusion-model-simulation-and-estimation
**Branch with paper.md** (empty if default branch):
**Version:** v2.0
