### 2023-10-02 #5762 @bkrayfield
Hello @samhforbes, I completed at least one round of review on the repository. I opened any concerns I had with the software repository directly. It has only been a week since those issues were raised. How long should I wait to see if they are addressed before completing my review?
