### 2024-04-11 #5762 @taqtiqa-mark
@samhforbes , from my PoV I'd expect all issues to be addressed.  I'm particularly concerned about the inability or reluctance to reproduce some figures from one of the original papers - of course it is possible the original paper is incorrect. However, I have directed the authors to one existing replication of one of the figures at issue, so I am doubtful that will be the case.  Hopefully we find out.
