### 2024-04-12 #5762 @peilun-he
Hi @samhforbes , yes we are working on a new version about unit testing, docker installation, and the two requried figures. Also we will be making some changes to the UI to incorporate the unit testing, which will take some time. All these should be done by the end of April and hopefully this version will solve all issues. Sorry for the delay.
