### 2024-05-27 #5762 @taqtiqa-mark
@samhforbes, I have spent considerable time and effort to try and assist the authors revise their submission. Previously [I asked](https://github.com/peilun-he/PDSim/issues/21#issuecomment-2016604305):

> Please only update these issues when you believe you have implemented what I have suggested is need. Or ask an editor to disregard the suggestions.

Consequently, I have taken the authors update as their final response.

Unfortunately, I must recommend their submission be rejected and they not be invited to resubmit it, nor any related work.

There are numerous reasons for this. However, I will address one that, to my mind, is determinate.
Namely, their persistent refusal to properly cite the source of their ideas, and properly attribute the work of others.
I have raised this issue to little effect:
 - [here](https://github.com/peilun-he/PDSim/issues/14#issuecomment-1937402198):
> You keep referring to "our model" - is this not just the FP(sic) model? If not please clarify how "your" model differs from FP(sic).

Yet the authors persist in making the claim they do more than (badly) implement the model of Filipovic & Larsen, they continue to claim some degree of authorship:

- https://github.com/peilun-he/PDSim/blob/66939da6a1bf4de325486a2bc8adea4665a33289/README.md?plain=1#L336-L338

Related, as I noted [previously](https://github.com/openjournals/joss-reviews/issues/5762#issuecomment-2049987966), the authors steadfastly refused to replicate some existing results, see [here](https://github.com/peilun-he/PDSim/issues/14), [here](https://github.com/peilun-he/PDSim/issues/14#issuecomment-1953263768), [here](https://github.com/peilun-he/PDSim/issues/14#issuecomment-2025776263), [here](https://github.com/peilun-he/PDSim/commit/091cbeb9c9f803c15cf613ee552c91b2d499741f#commitcomment-138492536) and 3-months later [here](https://github.com/peilun-he/PDSim/issues/4#issuecomment-2016598045), .
Only doing so when it was brought to your attention as an issue. As I noted at the time [I had to find](https://github.com/peilun-he/PDSim/issues/21#issuecomment-2016604305) an existing replication, [despite earlier asking them this question directly](https://github.com/peilun-he/PDSim/commit/091cbeb9c9f803c15cf613ee552c91b2d499741f#commitcomment-138817619), and ask they at least mimic that authors calculations. They were [incapable of even that level of effort](https://github.com/peilun-he/PDSim/issues/21#issuecomment-2109052113).

To make matters worse, they have (again) seen fit to refuse to acknowledge that prior art (replication) they are mimicing. While there is a reference in the readme file, nowhere in the body of their work do they cite and acknowledge [the work](https://au.mathworks.com/matlabcentral/fileexchange/29745-simulation-of-schwartz-smith-two-factor-model?s_tid=FX_rc3_behav) of [Moeti Ncube](https://au.mathworks.com/matlabcentral/profile/authors/2468914), and the role it played in helping them resolve their profound confusion about the objects they have been dealing with.

As I indicated above there remain substantial defects in the current submission. Defects I am confident they cannot remedy without substantial assistance from myself or one similarly well versed in the art. However, it is the fact the authors continue to hold out the work of Filipovic & Larsen as "our model", and their choice not to acknowledge the prior art of M. Ncube, which lead me to recommend the submission be rejected and the authors not be invited to resubmit it, nor any related work.

I understand the editors may review the facts and arrive at a different conclusion.
I acknowledge that reasonable minds can disagree on these points, and respect their right to do so.
I that case, I simply ask the editors remove my name as a reviewer available to the journal.

Sincerely
Mark Van de Vyver
