### 2024-06-05 #5762 @peilun-he
Dear Editor,

First of all, thank you for keeping access to the repository open and enabling us to respond to Dr Van de Vyver's comments. For us, the paper review process through open dialogue has been a highly stimulating and positive experience until this very moment, regardless of the outcome. In appreciation of the reviewers' time and efforts dedicated to improving our work, we have added the [Acknowledgments](https://github.com/peilun-he/PDSim/blob/f52fa74f5aa8b6e8d23aa7baf3ba517a03871f8b/README.md?plain=1#L805C1-L814C71) section to the paper. Furthermore, we will address the comments made by the Reviewer:

### 1.

> Namely, their persistent refusal to properly cite the source of their ideas, and properly attribute the work of others.

We have updated the [references](https://github.com/peilun-he/PDSim/blob/f52fa74f5aa8b6e8d23aa7baf3ba517a03871f8b/README.md?plain=1#L342C50-L345C37) according to the Reviewer's suggestion about the previous work done by Ncube (2010) on the Schwartz-Smith (2000) model. Now, we have also cited the work by Goodwin (2015) and He (2020). Therefore, thanks to the Reviewer's comments, we have included the above three references in the README.md file. We have also included references to the hyperlinks for Ncube, Goodwin and He's code in the paper.

> You keep referring to "our model" - is this not just the FP(sic) model? If not please clarify how "your" model differs from FP(sic).
> Yet the authors persist in making the claim they do more than (badly) implement the model of Filipovic & Larsen, they continue to claim some degree of authorship: [https://github.com/peilun-he/PDSim/blob/66939da6a1bf4de325486a2bc8adea4665a33289/README.md?plain=1#L336-L338](https://url.au.m.mimecastprotect.com/s/qENiCxngGkf7zOkyF8N-FV?domain=aus01.safelinks.protection.outlook.com)

To clarify immediately, we have never claimed ownership of Filipovic and Larsson's model. Indeed, quoting from the above link:

"... the parameter $\gamma$ is set to zero. However, **in our extended model**, we introduce the flexibility for this mean-reversion parameter associated with the long-term factor to take on non-zero values. Arbitrage-free pricing..."

**"our"** only appeared once, where it was used in reference to the extended Schwartz-Smith model with $\gamma \ge 0$ as opposed to the original Schwartz and Smith (2000) model with $\gamma = 0$.

We have rectified this error by completely rephrasing the relevant paragraph. The paragraph (containing the quote with "our extended model" above) now reads as [this](https://github.com/peilun-he/PDSim/blob/f52fa74f5aa8b6e8d23aa7baf3ba517a03871f8b/README.md?plain=1#L336C1-L342C49).

### 2.

> Related, as I noted [previously](https://url.au.m.mimecastprotect.com/s/0T-fCyoj8Pu8pnZgiMh3xw?domain=aus01.safelinks.protection.outlook.com), the authors steadfastly refused to replicate some existing results, see [here](https://url.au.m.mimecastprotect.com/s/WR5_Czvkmpfk7L63cgsKxN?domain=aus01.safelinks.protection.outlook.com), [here](https://url.au.m.mimecastprotect.com/s/mXcdCANpnDCw8Y4gcY40_P?domain=aus01.safelinks.protection.outlook.com), [here](https://url.au.m.mimecastprotect.com/s/JHTlCBNqgBCwyoj1cWpq-p?domain=aus01.safelinks.protection.outlook.com), [here](https://url.au.m.mimecastprotect.com/s/RMk_CD1vRkCEyPq1FkGl1F?domain=aus01.safelinks.protection.outlook.com) and 3-months later [here](https://url.au.m.mimecastprotect.com/s/uTLrCE8wlRCvYPLmfBszaX?domain=aus01.safelinks.protection.outlook.com).

These two figures depict the paths of the futures' price processes simulated using [Filipovic and Larsson ](https://github.com/peilun-he/PDSim/blob/main/figures/PD_Est_Futures.png)(Polynomial Diffusion), and [Schwartz and Smith](https://github.com/peilun-he/PDSim/blob/main/figures/SS_Est_Futures.png)'s models.  After this, in approximately 3 months, we reproduced [Figure 1](https://github.com/peilun-he/PDSim/blob/main/figures/SS_Figure1.png) from Schwartz and Smith's (2000) paper, using the estimates from this paper. The graph is titled 'Simulated Spot Price and long-term component (SS model)' and can be found under "Replicating Schwartz and Smith's Results" section. We have also produced the [graph](https://github.com/peilun-he/PDSim/blob/main/figures/PD_Figure1.png) for 'Simulated Spot Price and long-term component (PD model)' based on Filipovic and Larsson's model. For comparison, this  graph can also be found  under "Replicating Schwartz and Smith's Results" section. Further, the [joint plot](https://github.com/peilun-he/PDSim/blob/main/figures/real_data.png) of the Simulated Futures Prices obtained from the SS and PD models can be found in the section "Simulation Accuracy".
We remain very thankful to the Reviewer for pointing out that the "PDSim" had a fault, it was unable to simulate the paths of spot prices from the original Schwartz-Smith's model with $\gamma=0$ due to indefiniteness caused by division on zero in the code; this problem has been rectified since then by adding a new function, specifically designed to handle this scenario.
Further, thanks to the reference to Ncube (2010) shared by the Reviewer, we were able to download the data provided by the author on Matlab File Exchange.  We sincerely apologise for not noticing the folder with the data earlier; we have now been able to replicate the following graph from Ncube (2010). The required estimates are obtained using the original SS(2000) model with $\gamma=0$  and the [graph](https://github.com/peilun-he/PDSim/blob/main/figures/Replicated_figure4.png) "Estimated spot price and equilibrium price" that can be viewed under "Replicating Schwartz and Smith's Results" section. The R code for producing this graph  is given [here](https://url.au.m.mimecastprotect.com/s/suipCk815RCKyn0ncVsHW_?domain=aus01.safelinks.protection.outlook.com).

### 3.

> As I noted at the time [I had to find](https://url.au.m.mimecastprotect.com/s/P825CwV1jpSq2D8AhVh0SX?domain=aus01.safelinks.protection.outlook.com) an existing replication, [despite earlier asking them this question directly](https://url.au.m.mimecastprotect.com/s/zc_DCxngGkf7z2M9IvioOu?domain=aus01.safelinks.protection.outlook.com), and ask they at least mimic that authors calculations. They were [incapable of even that level of effort](https://url.au.m.mimecastprotect.com/s/LdJHCyoj8Pu8plX0hQH4CE?domain=aus01.safelinks.protection.outlook.com).

As we addressed, in the reply to 2, using the data from Ncube (2010) and the parameters estimates from Schwartz and Smith (2000), we replicated the figure from Schwartz and Smith (2000) with ($\gamma=0$) for the Estimated Spot Price and Equilibrium Spot Price. This [graph](https://github.com/peilun-he/PDSim/blob/main/figures/Replicated_figure4.png) is included in the section "Replicating Schwartz and Smith's Results" section.

### 4.

> To make matters worse, they have (again) seen fit to refuse to acknowledge that prior art (replication) they are mimicing. While there is a reference in the readme file, nowhere in the body of their work do they cite and acknowledge [the work](https://url.au.m.mimecastprotect.com/s/nxcnCzvkmpfk70gxsKQuVl?domain=au.mathworks.com) of [Moeti Ncube](https://url.au.m.mimecastprotect.com/s/R9osCANpnDCw8v6nhQ3_Jd?domain=au.mathworks.com), and the role it played in helping them resolve their profound confusion about the objects they have been dealing with.

As we indicated above, we have now cited Ncube (2010)'s work in reference to Schwartz-Smith's original model, where the long-term equilibrium process represents a geometric Brownian motion (gBm). Additionally, we have included other references related to the implementation of algorithms for parameter estimation of Schwartz-Smith's extended model with $\gamma \ge 0$, where the long-term equilibrium process is an Ornstein-Uhlenbeck process. The relevant paragraph is [here](https://github.com/peilun-he/PDSim/blob/818f383adaae7470451edc4c1f5a9359377e09e2/README.md?plain=1#L342C50-L346C1).

Kind regards,
Peilun He, Nino Kordzakhia, Gareth Peters and Pavel Schevchenko
