### 2024-08-01 #5762 @samhforbes
Hi @peilun-he I appreciate your patience. This submission has been discussed with various editors and EiCs at JOSS.
In this case I’m afraid we are going to have to reject this manuscript at JOSS on the back of this review procedure.
There’s clearly potential in this software, and given this has been a very long review I hope the feedback has been useful in improving the work. I would recommend submitting the work to another software journal in this case, and sincerely wish you the best of luck with it.
Thank you to both reviewers for a thorough job.
